fix: Include 'grapheme' JSON file for 'alive-progress' compatibility
Commit addresses an issue, where the 'alive-progress' library relies on the 'grapheme' library, which requires a specific JSON file. The JSON file was not included by default by PyInstaller during the build process. This fix dynamically finds the installation path of 'grapheme' and includes the necessary JSON file required by 'grapheme', which is in turn used by 'alive-progress'.