feat: Enhance Workload Management with Bulk Deletion and 'All' Flag Support
* feat: add support for bulk deletion and introduce 'all' flag in workload management

Enhanced the workload management CLI tool by adding the capability to pass a list of workloads for deletion. This update allows users to specify multiple workloads simultaneously, improving the deletion process's efficiency. The introduction of an 'all' flag enables the deletion of all workloads, enhancing the tool's flexibility and efficiency in managing GitOps workloads.

---------

Co-authored-by: Alexander Ulyanov <alexu@cloudgeometry.io>